pyproject.toml
[tool.poetry]
name = "spotify-backend"
version = "0.1.0"
description = ""
authors = ["dinmukhammed <arystambekdimash005@gmail.com>"]
readme = "README.md"
packages = [{include = "spotify_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.23.2"
fastapi = "^0.100.1"
sqlalchemy = "^2.0.19"
pydantic = "^2.1.1"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
pillow = "^10.0.0"
aiofiles = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
